+++ test/bootstrap.php
@@ -14,7 +14,7 @@ if (!$loader = @include __DIR__ . '/../vendor/autoload.php') {
             'php composer.phar install' . PHP_EOL);
 }
 
-ini_set('display_errors', 1);
+ini_set('display_errors', 'true');
 ini_set('error_reporting', E_ALL);
 
 if (defined('HHVM_VERSION')) {