+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/) 
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## 2.2.3 (2017-08-17)
+
+### Improved
+
+- Now `Adapter` defaults to `pjb62` driver.  
+- Unit tests, preliminary work to update to phpunit 6+
+
 ## 2.2.2 (2017-08-17)
 
 ### Fixed